v1.1.0 * Fix typings
