--- build.config.mjs
@@ -1,36 +1,0 @@
-import {execSync} from "child_process";
-import {build} from "esbuild";
-import fs from "fs/promises";
-
-async function doBuild() {
-    const isProd = process.env.NODE_ENV === "production";
-
-    execSync("tsc -p tsconfig.esm.json", {stdio: "inherit"});
-
-    // Duplicate the types index, to have all the typings recognized correctly
-    await fs.copyFile("dist/types/index.d.ts", "dist/types/index.d.cts");
-
-    await build({
-        entryPoints: ["dist/raw-esm/**/*.js"],
-        outdir: "dist/esm",
-        format: "esm",
-        bundle: !1,
-        minify: !1 // never minify as it's only a source to work on
-    });
-
-    await build({
-        entryPoints: ["dist/raw-esm/index.js"],
-        outfile: "dist/cjs/index.cjs",
-        format: "cjs",
-        bundle: !0,
-        minify: isProd
-    });
-
-}
-
-try {
-    doBuild();
-}
-catch (e) {
-    console.error("Error");
-}
