+++ package.json
@@ -1,25 +1,25 @@
 {
     "name": "@tonysamperi/crypto",
-    "version": "1.0.0",
+    "version": "1.1.0",
     "author": "Tony Samperi",
     "description": "Modern tree-shakeable TypeScript version of crypto-js",
     "type": "module",
-    "typings": "./dist/types/index.d.cts",
+    "types": "./dist/esm/index.d.cts",
     "main": "./dist/cjs/index.cjs",
     "module": "./dist/esm/index.js",
     "unpkg": "./dist/cjs/index.cjs",
     "exports": {
         ".": {
             "import": {
-                "types": "./dist/types/index.d.ts",
+                "types": "./dist/esm/index.d.ts",
                 "default": "./dist/esm/index.js"
             },
             "require": {
-                "types": "./dist/types/index.d.cts",
+                "types": "./dist/cjs/index.d.cts",
                 "default": "./dist/cjs/index.cjs"
             },
             "browser": {
-                "types": "./dist/types/index.d.ts",
+                "types": "./dist/esm/index.d.ts",
                 "default": "./dist/esm/index.js"
             }
         },
@@ -28,14 +28,15 @@
     "sideEffects": false,
     "files": [
         "dist/**",
-        "!dist/raw-esm/**",
+        "!dist/raw-esm",
+        "!dist/raw-types",
         "README.md"
     ],
     "engines": {
         "node": ">=18"
     },
     "scripts": {
-        "build": "npm run clean && NODE_ENV=production node build.config.mjs",
+        "build": "npm run clean && NODE_ENV=production node build.config.js",
         "clean": "rimraf dist",
         "lint": "eslint && tsc --noEmit",
         "lint:fix": "eslint --fix",
@@ -63,7 +64,9 @@
         "esbuild": "^0.25.1",
         "eslint": "^9.22.0",
         "eslint-plugin-import": "^2.31.0",
+        "fast-glob": "^3.3.3",
         "jest": "^29.6.4",
+        "replace-in-file": "^8.3.0",
         "rimraf": "^5.0.7",
         "ts-jest": "^29.1.1",
         "ts-node": "^10.9.2",
